Allow custom storage object to be used
The strings 'cookies' and 'localStorage' are still allowed, but instead
of switching on the string for functionality, the string is replaced
by an object upon initialization, and the current methods just call
the matching methods on those objects. This allows a user to, instead
of specifying a string, set the storage to an object that implements
these methods, which allows them to store the auth data in literally
any way imaginable.